post method successful

+++ emp_app/models.py
@@ -20,10 +20,10 @@ class Employee(models.Model):
     first_name = models.CharField(max_length=200)
     last_name = models.CharField(max_length=200)
     dept = models.ForeignKey(Department, on_delete=models.CASCADE)
-    salary = models.IntegerField(default=0)
-    bonus = models.IntegerField(default=0)
+    salary = models.BigIntegerField(default=0)
+    bonus = models.BigIntegerField(default=0)
     role = models.ForeignKey(Role, on_delete=models.CASCADE)
-    phone = models.IntegerField(default=0)
+    phone = models.BigIntegerField(default=0)
     hire_date = models.DateField()
 
     def __str__(self):
